+++ examples/testmodule/plugins/__init__.py
@@ -6,10 +6,16 @@
 import os
 from pathlib import Path
 
+import inmanta_plugins.std as std
+from inmanta.agent.handler import CRUDHandler, HandlerContext, provider
+from inmanta.plugins import PluginException, plugin
+from inmanta.resources import PurgeableResource, resource
 
-from inmanta.agent.handler import provider, CRUDHandler, HandlerContext
-from inmanta.resources import resource, PurgeableResource
-from inmanta.plugins import plugin
+# perform side effect on another module in the inmanta_plugins package
+# to make sure it is only executed once (test_49_plugin_load_side_effects)
+if not hasattr(std, "pytest_inmanta_side_effect_count"):
+    std.pytest_inmanta_side_effect_count = 0
+std.pytest_inmanta_side_effect_count += 1
 
 
 @resource("testmodule::Resource", agent="agent", id_attribute="name")
@@ -27,7 +33,9 @@ class ResourceHandler(CRUDHandler):
     def create_resource(self, ctx: HandlerContext, resource: ResourceResource) -> None:
         ctx.set_created()
 
-    def update_resource(self, ctx: HandlerContext, changes: dict, resource: ResourceResource) -> None:
+    def update_resource(
+        self, ctx: HandlerContext, changes: dict, resource: ResourceResource
+    ) -> None:
         ctx.set_updated()
 
     def facts(self, ctx, resource):
@@ -43,7 +51,7 @@ class ResourceHandler(ResourceHandler):
     def read_resource(self, ctx: HandlerContext, resource: ResourceResource) -> None:
         resource.purged = False
         resource.value = "read"
-        ctx.warning("argument can not be serialized", argument={"a":"b"}.values())
+        ctx.warning("argument can not be serialized", argument={"a": "b"}.values())
 
 
 @plugin
@@ -53,3 +61,17 @@ def create_testfile():
 
 def regular_function():
     return "imported"
+
+
+class TestException(PluginException):
+    pass
+
+
+@plugin
+def get_exception():
+    return TestException
+
+
+@plugin
+def raise_exception():
+    raise TestException("my exception")