+++ examples/testmodule/tests/test_capture.py
@@ -4,6 +4,7 @@
     License: Apache 2.0
 """
 
+
 def test_capture(project):
     marker = "MARKERXXX123678"