+++ examples/testmodule/tests/test_project_no_plugins.py
@@ -1,0 +1,9 @@
+"""
+    Copyright 2020 Inmanta
+    Contact: code@inmanta.com
+    License: Apache 2.0
+"""
+
+
+def test_project_no_plugins_deprecation_warning(project_no_plugins) -> None:
+    pass