+++ pytest_inmanta/module/init.py
@@ -18,10 +18,9 @@
 
 from inmanta import resources
 from inmanta.agent import handler
-
 from pytest_inmanta.handler import DATA
 
-KEY_PREFIX="unittest_"
+KEY_PREFIX = "unittest_"
 
 
 @resources.resource("unittest::Resource", id_attribute="name", agent="agent")
@@ -31,8 +30,14 @@ class Resource(resources.PurgeableResource):
 
 @handler.provider("unittest::Resource", name="test")
 class ResourceHandler(handler.CRUDHandler):
-    def read_resource(self, ctx: handler.HandlerContext, resource: resources.PurgeableResource) -> None:
-        ctx.info("Resource fail %(fail)s skip %(skip)s", fail=resource.fail, skip=resource.skip)
+    def read_resource(
+        self, ctx: handler.HandlerContext, resource: resources.PurgeableResource
+    ) -> None:
+        ctx.info(
+            "Resource fail %(fail)s skip %(skip)s",
+            fail=resource.fail,
+            skip=resource.skip,
+        )
 
         if resource.skip:
             raise handler.SkipResource()
@@ -54,18 +59,41 @@ class ResourceHandler(handler.CRUDHandler):
 
         resource.desired_value = DATA[resource.name]["desired_value"]
 
-    def create_resource(self, ctx: handler.HandlerContext, resource: resources.PurgeableResource) -> None:
+    def create_resource(
+        self, ctx: handler.HandlerContext, resource: resources.PurgeableResource
+    ) -> None:
         DATA[resource.name] = {}
         DATA[resource.name]["desired_value"] = resource.desired_value
 
         ctx.set_created()
 
-    def delete_resource(self, ctx: handler.HandlerContext, resource: resources.PurgeableResource) -> None:
+    def delete_resource(
+        self, ctx: handler.HandlerContext, resource: resources.PurgeableResource
+    ) -> None:
         del DATA[resource.name]
 
         ctx.set_purged()
 
-    def update_resource(self, ctx: handler.HandlerContext, changes: dict, resource: resources.PurgeableResource) -> None:
+    def update_resource(
+        self,
+        ctx: handler.HandlerContext,
+        changes: dict,
+        resource: resources.PurgeableResource,
+    ) -> None:
         DATA[resource.name]["desired_value"] = resource.desired_value
 
         ctx.set_updated()
+
+
+@resources.resource("unittest::IgnoreResource", id_attribute="name", agent="agent")
+class IgnoreResource(resources.PurgeableResource):
+    fields = ("name", "desired_value")
+
+    @staticmethod
+    def get_desired_value(exporter, resource):
+        raise resources.IgnoreResourceException()
+
+
+@handler.provider("unittest::IgnoreResource", name="ignore")
+class IgnoreResourceHandler(handler.CRUDHandler):
+    pass